(i don't like this!) Dictionary as a non-ssr component
